Update addresses and signatures for 5.1.

+++ Dalamud/Game/ClientState/Structs/JobGauge/DRKGauge.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 namespace Dalamud.Game.ClientState.Structs.JobGauge {
     [StructLayout(LayoutKind.Explicit)]
     public struct DRKGauge {
-        [FieldOffset(0)] public short Blood;
+        [FieldOffset(0)] public byte Blood;
         [FieldOffset(2)] public short DarksideTimeRemaining;
         [FieldOffset(4)] private byte DarkArtsState;
         [FieldOffset(6)] public short ShadowTimeRemaining;
